Fix AVAudioEngine deadlock, CPU hotpath, and misc improvements
Fixes:
- StreamingAudioPlayer: cache volume in _cachedVolume to avoid accessing
  mainMixerNode from inside tap callback — prevents three-way mutex deadlock
  during streaming crossfade (Thread A: stopEngine, B: volume setter, C: tap)
- PlexBrowserView: skip tab/list redraw when dirtyRect is in server bar zone
  only; prevents 15Hz server scroll timer from triggering expensive drawListArea
  / drawColumnRow CoreText rendering on every tick
- VisClassicCore: fix right-channel bar loop starting at i=half instead of i=0
- RadioManager: handle IPv6 bracket notation and add IPv6 loopback/link-local/
  ULA detection in isPrivateHost

Modern skin:
- window.opacity is now optional (defaults to 1.0); custom init(from:) uses
  decodeIfPresent — old skins without the key load correctly
- ModernSkinEngine: try exact skin name before resolved name when loading
- ModernSkinAnimation: use RunLoop.main.add(.common) so timer fires during menus

vis_classic:
- Add VolumeFunction INI section read/write support
- Upstream LevelCalc/LevelFunc/LogBarTable/BarColour: volumeFunc array

Other:
- MediaLibrary: scanGeneration counter to discard stale in-flight scan results
- JellyfinServerClient: improve slow-request and pagination log messages
- PlexServerClient: redact X-Plex-Token in debug smart-playlist URL log
- Info.plist: bump build number to 2

+++ Sources/CVisClassicCore/upstream/LevelFunc.cpp
@@ -63,5 +63,5 @@ void SqrTable(int *table)
 void FractionalTable(int *table)
 {
   for(int i = 0; i < 256; i++)
-    table[i] = (int)(256.0 - (double)(256 / (table[i] + 1)));  // 1/x function
+    table[i] = (int)(256.0 - 256.0 / (table[i] + 1));  // 1/x function
 }
